feat: 拆分并完善尺寸相关迁移逻辑
