+++ schematics/ng-update/update-22/index.ts
@@ -1,10 +1,12 @@
 import { Rule } from '@angular-devkit/schematics';
 import { createMigrationSchematicRule, NullableDevkitMigration, TargetVersion } from '@angular/cdk/schematics';
+import { ClassNamesMigration } from './class-name-migration';
 import { onMigrationComplete } from '../core/complete';
 import { InputControlSizeMigration } from './input-control-size-migration';
+import { SassSizeVariableMigration } from './sass-size-variable-migration';
 import { upgradeData } from './update-data';
 
-const migrations: NullableDevkitMigration[] = [InputControlSizeMigration];
+const migrations: NullableDevkitMigration[] = [InputControlSizeMigration, SassSizeVariableMigration, ClassNamesMigration];
 
 export default function main(): Rule {
     return createMigrationSchematicRule(TargetVersion.V22, migrations, upgradeData, onMigrationComplete);
